+++ lib/shoe.rb
