+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
-require_relative '../book'
-require_relative '../shoe'
+require_relative '../lib/book'
+require_relative '../lib/shoe'
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
